rename location to default instead of route as some middlewares are http and other route middlewares
