+++ src/masonite/middleware/default/CorsMiddleware.py
