+++ src/masonite/middleware/default/EncryptCookies.py
