+++ src/masonite/middleware/default/GuardMiddleware.py
