+++ src/masonite/middleware/default/LoadUserMiddleware.py
